badges now also use the refit api
